Fix bug in schollu
