+++ scholly/base/src/components/publications/Publications.js
@@ -1,26 +1,57 @@
 /**
  * The Publications component displays a list of publications.
  */
-import React, { Fragment, useState } from 'react';
+import React, { Fragment, useState, useEffect } from 'react';
 import { Accordion } from 'react-bootstrap';
 import LayoutAllPublications from './publicationsLayout/LayoutAllPublications';
 import LayoutByCategory from './publicationsLayout/LayoutByCategory'
 import { TEAM_PUBLICATIONS } from '../../global/data';
 import PublicationsDropdown from './publicationsLayout/PublicationsDropdown';
 import { WEB_PAGES } from '../../global/data';
-import { layoutOption } from '../../config/publications';
+import { layoutOption, sortingOption } from '../../config/publications';
 
 const Publications = () => {
   const { publicationOptions } = WEB_PAGES
-  console.log(WEB_PAGES)
-  console.log(publicationOptions)
-  const [preference, setPreference] = useState(publicationOptions);
+  const [ options, setOptions ] = useState(publicationOptions);
 
+  const sortPublications = (teamPublications, option) => {
+    switch (option) {
+        case sortingOption.AUTHOR:
+            teamPublications.sort((a, b) =>
+                a.authors[0].toLowerCase() > b.authors[0].toLowerCase() ? 1 : -1
+            );
+            break;
+        case sortingOption.TITLE:
+            // publication title
+            teamPublications.sort((a, b) =>
+                a.title.toLowerCase() > b.title.toLowerCase() ? 1 : -1
+            );
+            break;
+        case 'Category Title':
+            // journal or conference title
+            teamPublications.sort((a, b) =>
+                a.category.categoryTitle.toLowerCase() >
+                b.category.categoryTitle.toLowerCase()
+                ? 1
+                : -1
+            );
+            break;
+        default:
+            // sort by title then year for consistency with the db
+            teamPublications.sort((a, b) =>
+                a.title.toLowerCase() > b.title.toLowerCase() ? 1 : -1
+            );
+            teamPublications.sort((a, b) => (a.year > b.year ? -1 : 1));
+            break;
+        }
+        return teamPublications
+    };
+    
+  const [ publications, setPublications ] = useState(sortPublications(TEAM_PUBLICATIONS, options.sortBy));
 
-  
   const renderPublications = () => {
-    console.log(preference)
-    switch (preference.layout) {
+    console.log(options)
+    switch (options.layout) {
       case layoutOption.BY_CATEGORY:
         return <LayoutByCategory teamPublications={TEAM_PUBLICATIONS}/>;
       default:
@@ -28,12 +59,14 @@ const Publications = () => {
     }
   };
 
+
   return (
     <Fragment>
       <PublicationsDropdown 
-        preference={preference}
-        setPreference={setPreference}
-        publication={TEAM_PUBLICATIONS}
+        options={options}
+        setOptions={setOptions}
+        publications={publications}
+        sortPublications={sortPublications}
       />
       <Accordion>
         {renderPublications()}
